Thermostat It started from an ideea of David Bird
This code aims to make your device work even when there is no internet connection. In this situation, the device will actually turn on only after a date and current time have been set for it to be able to function according to the schedules made. In order not to have problems with maintaining the date and time of the device, it would be good for it to be protected by using a battery for power or an RTC module (need suplimentary implementations). After restarting the device without internet access, it needs to set the date and time. It is enough to access it as an Access Point and it will ask you to enter the current date and time.